Refactor `worldSpace` `vec2` -> `ivec2`
Changed all 'pixel dependent logic', like the world size from `glm::vec2` to `glm::ivec2`

+++ lib/include/chomper/player.hpp
@@ -36,8 +36,8 @@ class Player : public IController::IListener, public IDebugInspector, public kli
 	}
 
   private:
-	[[nodiscard]] bool isCollidingWithSelf(glm::vec2 targetGrid) const;
-	[[nodiscard]] bool isCollidingWithWall(glm::vec2 targetGrid) const;
+	[[nodiscard]] bool isCollidingWithSelf(glm::ivec2 targetGrid) const;
+	[[nodiscard]] bool isCollidingWithWall(glm::ivec2 targetGrid) const;
 	void move();
 	void updateScoreText();
